[refactor] move QRCodeScanner/ to become a separate tab bar item.
