+++ src/app/(BottomTabNavigation)/QRCodeScanner/_layout.tsx
@@ -1,0 +1,20 @@
+import { Stack } from 'expo-router';
+
+import { HeaderImage } from '../../../Components/HeaderComponents/HeaderComponents';
+
+export default function QRCodeScannerLayout() {
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: true,
+      }}
+    >
+      <Stack.Screen
+        name="index"
+        options={{
+          headerTitle: () => <HeaderImage />,
+        }}
+      />
+    </Stack>
+  );
+}
