--- src/app/(BottomTabNavigation)/Updates/_layout.tsx
@@ -1,6 +1,0 @@
-import { Stack } from 'expo-router';
-import React from 'react';
-
-export default function ProfileLayout() {
-  return <Stack screenOptions={{ headerShown: false }} />;
-}
