--- src/app/(BottomTabNavigation)/Updates/index.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Text, View } from 'react-native';
-
-import styles from './styles';
-
-function ProfileScreen() {
-  return (
-    <View style={styles.container}>
-      <Text>This is the /Updates/index update screen!</Text>
-    </View>
-  );
-}
-
-export default ProfileScreen;
