--- src/app/(BottomTabNavigation)/Updates/styles.ts
@@ -1,12 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-import { colors } from '../../../styles/colors';
-
-export default StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: colors.white,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
